# Conflicts:
#	St-dafint.ab/src/main/resources/application.properties

+++ St-dafint.ab/src/main/java/com/example/nicecleaning/dto/CleanResponseDTO.java
@@ -11,6 +11,6 @@ public record CleanResponseDTO(
         // Statusen på den. Bekräftad, obekräftad, osv.
         int status,
         // Vem som bokade den. Ska bytas ut mot vem städaren är i slutändan
-        int appUserId) {
+        Long appUserId) {
 
 }
